requirements.txt
fire
speedtest-cli
prettyTable
pyinstaller